MPVActivity: fix malfunctioning PiP controls
Apparently PiP used to imply multi-window mode, but it no longer does.
This poses a second problem however: when exiting PiP Android will
inform us of the state change while Activity getters claim
that we're not in PiP but in multi-window mode, breaking more assumptions.
(does this make any sense? no it does not)
So add some code to hack around that too.