MPVView: default to OpenGL ES
This used to be less important. Now we have the no-copy hwdec
which only works with ES plus mpv's fallback to copy doesn't
seem to work (though that's a bug for another time).
Effectively I think the NVIDIA Shield is the only device that even
supported desktop OpenGL and will be affected.

Note that opengl-es=yes actually forces GLES and won't fall back,
GLES is an install requirement for mpv-android so this won't be a problem.