app/build: raise version for universal apk/bundle by 8000
Since we're switching from arch-specific APKs to an app bundle
(for Google Play distributions) existing users need an upgrade path.
Existing arch-specific APKs have a versionCode in the range
1000 to 4000(*), which is higher than the universal one at 31.
So now we have to do this bullshit to make this work.

(*): Have I mentioned that it's only this way due to GP requirements?