MPVActivity: close file descriptors when no longer needed
3ada286e34625068bd74370db7cf413614e5c7df adjusted things so that fd://
URIs are used instead of fdclose:// URIs due to the latter not being
reusable when using playlists. However, this left no mechanism to close
file descriptors that are no longer needed, which can cause SAF content
providers to keep running in the background indefinitely.

This commit updates MPVActivity's logic to keep track of open file
descriptors and close them when they are no longer needed. This is
primarily when MPVActivity is destroyed.